base de donnee comune

+++ mat.h
@@ -8,20 +8,21 @@ class Mat
 {
 public:
     Mat();
-    Mat(int,int,float,QString,QString,QString);
+    Mat(int,int,int,QString,QString,QString);
     int getnums();
     int getquantiter();
-    float getprix();
+    int getprix();
     QString getnomf();
     QString getnomm();
     QString gettype();
 
     void setnums(int);
     void setquantiter(int);
-    void setprix(float);
+    void setprix(int);
     void setnomf(QString);
     void setnomm(QString);
     void settype(QString);
+
     bool ajouter_materiel();
     QSqlQueryModel* afficher_materiel();
     bool supprimer_materiel(int);
@@ -32,7 +33,7 @@ public:
 private:
     int num_serie;
     int quantiter;
-    float prix;
+    int prix;
     QString nom_fournisseur;
     QString nom_materiel;
     QString type;
